news-summary/requirements.txt
requests
beautifulsoup4
slack_sdk
transformers
torch
sentencepiece
python-dotenv
